llm/pyproject.toml
[project]
name = "fraud-detection-llm"
version = "0.1.0"
description = "FastAPI backend for Fraud Detection ML models"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "networkx>=3.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
]

[project.scripts]
start = "uvicorn main:app --reload --port 8000"
